text-frontend: Don't crash when printer doesn't exist
Instead of crashing due to a NULL dereference,
print a message and skip further processing of
the current command when `cpdbFindPrinterObj`
returns NULL, e.g. because the user has entered
the name of a nonexisting printer or backend.

For some commands, null checks were in place
without printing a message. Unify that.