text-frontend: Get locale via GLib
Instead of only supporting the LANGUAGE environment variable
that's not necessarily set, use GLib's `g_get_language_names` to
detect the locale to use in the text frontend which takes
other environment variables into account as well [1]:

> The returned list is sorted from most desirable to
> least desirable and always contains the default locale “C”.

Also, make `locale` a local variable in the control thread
that's the only user instead of using a static variable.

This e.g. makes

    > get-all-translations PDF CUPS

print the translations for the current language in a fresh
Ubuntu 23.10 installation where the `LANGUAGE` environment
variable is not set, and the result without this commit
in place was therefore just

    > get-all-translations PDF CUPS
    No translations found

[1] https://docs.gtk.org/glib/func.get_language_names.html