Fix memory leaks in cpdbUnpackOptions
Make sure that the memory allocated while
unpacking the variant is freed again, as
described in more detail in the `g_variant_get` [1]
and in the GVariant Format Strings doc [2].

For arrays, the iterator needs to be freed
using `g_variant_iter_free` as described at [3].

In general, let `g_variant_iter_loop` take care
of freeing the memory as described in [4].
However, add an additional check to detect
the case where a backend might be sending
invalid data and sends an array that contains
more elements than the reported amount. Free
memory manually in this case.
(No mismatch was observed with printers
reported by cpdb-backend-cups in my local tests with
additional asserts that the array element
count matches the reported amount of elements.)

This addresses memory leaks like the following reported by
valgrind, seen e.g. when running `cpdb-text-frontend` and running
the command `get-all-options PDF CUPS` to print the options of the
CUPS-PDF printer, then quitting:

    ==2670840== 50,392 (80 direct, 50,312 indirect) bytes in 1 blocks are definitely lost in loss record 1,411 of 1,412
    ==2670840==    at 0x4842808: malloc (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
    ==2670840==    by 0x4913B41: g_malloc (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x4955DA6: g_variant_iter_new (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x4957F32: ??? (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x4958CFD: g_variant_get_va (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x4958F72: g_variant_get (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x485DFC5: cpdbUnpackOptions (cpdb-frontend.c:2124)
    ==2670840==    by 0x485B399: cpdbGetAllOptions (cpdb-frontend.c:1062)
    ==2670840==    by 0x10ACEE: control_thread (cpdb-text-frontend.c:197)
    ==2670840==    by 0x493C2E0: ??? (in /usr/lib/x86_64-linux-gnu/libglib-2.0.so.0.8101.0)
    ==2670840==    by 0x4A8D6C1: start_thread (pthread_create.c:447)
    ==2670840==    by 0x4B07F6F: clone (clone.S:100)

(Valgrind still reports memory leaks elsewhere, but they're
unrelated to this change.)

While at it, also indent the involved code consistently
with spaces instead of a mixture of spaces and tabs.

[1] https://docs.gtk.org/glib/method.Variant.get.html
[2] https://docs.gtk.org/glib/gvariant-format-strings.html
[3] https://docs.gtk.org/glib/gvariant-format-strings.html#arrays
[4] https://docs.gtk.org/glib/method.VariantIter.loop.html