Drop unused cpdbExtractFileName
The way that `cpdbExtractFileName` casts a `const char*`
to a `char*` and returns a `char*` looks suspicious, as the
signature seems to suggest that it would be the caller's
responsibility to free the returned string, but it is
not actually the owner, as no new string was allocated.

In any case, `cpdbExtractFileName`, or more precisely
its predecessor `extract_file_name` is unused
since

    commit cde7d1dcf6fded0ec1147d87eaee5114e78b3c23
    Date:   Thu Aug 24 20:57:57 2017 +0530

        Remove the backend specific parts from original code

and the then already unused function was later renamed to
`cpdbExtractFileName` in

    commit b394eaa2369fe105bb4e28222456c0239110070f
    Date:   Sun Aug 14 09:41:25 2022 +0200

        Renamed all API functions, data types and constants

Since the function is neither used here in cpdb-libs nor in
cpdb-backend-cups, just drop it altogether.