Bump soname of both libraries from 2 to 3
Somewhere between 2.0b5 and 2.0b7 some API functions have been removed, assuming that while in beta, the API is not expected to be stabilized.

But as some distros have already included 2.0b5 we are bumping the soname, to not cause confusion with the distro's package build servers.

Solves PR #84 but we go the simpler way of just using "3" as new soname.